allow larger scenes

+++ src/osrs/painters.h
@@ -35,21 +35,21 @@ enum PaintersTileFlags
 struct PaintersTile
 {
     // These are only used for normal locs
-    uint16_t scenery[MAX_SCENERY_COUNT];
+    int32_t scenery[MAX_SCENERY_COUNT];
     uint16_t scenery_count;
 
-    int16_t wall_a;
-    int16_t wall_b;
-    int16_t wall_decor_a;
-    int16_t wall_decor_b;
+    int32_t wall_a;
+    int32_t wall_b;
+    int32_t wall_decor_a;
+    int32_t wall_decor_b;
 
-    int16_t ground_decor;
+    int32_t ground_decor;
 
     int32_t bridge_tile;
 
-    int16_t ground_object_bottom;
-    int16_t ground_object_middle;
-    int16_t ground_object_top;
+    int32_t ground_object_bottom;
+    int32_t ground_object_middle;
+    int32_t ground_object_top;
 
     // Contains directions for which tiles are waiting for us to draw.
     // This is determined by locs that are larger than 1x1.
